# mysql: insert entity and metadata inside of a transaction