Makes a pretty significant performance improvement when inserting an entity. On my system it's on the order of 0.3s without and 0.03s with a transaction.

Before:

```
PHP message: Known (known.dev): DEBUG - insert or update took 0.30136895179749 s
PHP message: Known (known.dev): DEBUG - insert or update took 0.3676438331604 s
PHP message: Known (known.dev): DEBUG - insert or update took 0.29871201515198 s
PHP message: Known (known.dev): DEBUG - insert or update took 0.37608504295349 s
```

After: 

```
PHP message: Known (known.dev): DEBUG - insert or update took 0.038892030715942 s
PHP message: Known (known.dev): DEBUG - insert or update took 0.024508953094482 s
PHP message: Known (known.dev): DEBUG - insert or update took 0.030456066131592 s
PHP message: Known (known.dev): DEBUG - insert or update took 0.024585962295532 s
```